+++ spec/spec_helper.rb
@@ -8,11 +8,3 @@ RSpec.configure do |config|
   config.filter_run_when_matching focus: true
   config.filter_run_excluding type: "e2e"
 end
-
-def with_term_on_empty(value)
-  old_value = ENV["TERM_ON_EMPTY"]
-  ENV["TERM_ON_EMPTY"] = value
-  yield
-ensure
-  ENV["TERM_ON_EMPTY"] = old_value
-end